+++ src/compiler/parser.c
@@ -9,7 +9,6 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <memory.h>
 #include <assert.h>
 #include <stdarg.h>